move locations to standalone file
