+++ vyper/semantics/data_locations.py
@@ -1,0 +1,9 @@
+import enum
+
+
+class DataLocation(enum.Enum):
+    UNSET = 0
+    MEMORY = 1
+    STORAGE = 2
+    CALLDATA = 3
+    CODE = 4
